requirements.txt
streamlit
cryptography
reedsolo
mysql-connector-python
